Rework security handler factory into sync and async
