+++ connexion/apis/flask_api.py
@@ -17,7 +17,7 @@ from connexion.apis.abstract import AbstractAPI
 from connexion.handlers import AuthErrorHandler
 from connexion.jsonifier import Jsonifier
 from connexion.lifecycle import ConnexionRequest, ConnexionResponse
-from connexion.security import FlaskSecurityHandlerFactory
+from connexion.security import SyncSecurityHandlerFactory
 from connexion.utils import is_json_mimetype, yamldumper
 
 logger = logging.getLogger('connexion.apis.flask_api')
@@ -28,7 +28,7 @@ class FlaskApi(AbstractAPI):
     @staticmethod
     def make_security_handler_factory(pass_context_arg_name):
         """ Create default SecurityHandlerFactory to create all security check handlers """
-        return FlaskSecurityHandlerFactory(pass_context_arg_name)
+        return SyncSecurityHandlerFactory(pass_context_arg_name)
 
     def _set_base_path(self, base_path):
         super()._set_base_path(base_path)
