+++ tests/conftest.py
@@ -1,11 +1,10 @@
 import json
 import logging
 import pathlib
-import sys
 
 import pytest
 from connexion import App
-from connexion.security import FlaskSecurityHandlerFactory
+from connexion.security import SyncSecurityHandlerFactory
 from werkzeug.test import Client, EnvironBuilder
 
 logging.basicConfig(level=logging.DEBUG)
@@ -90,12 +89,12 @@ def oauth_requests(monkeypatch):
                 return FakeResponse(200, '{"uid": "test-user", "scopes": ["myscope", "otherscope"]}')
         return url
 
-    monkeypatch.setattr('connexion.security.flask_security_handler_factory.session.get', fake_get)
+    monkeypatch.setattr('connexion.security.sync_security_handler_factory.session.get', fake_get)
 
 
 @pytest.fixture
 def security_handler_factory():
-    security_handler_factory = FlaskSecurityHandlerFactory(None)
+    security_handler_factory = SyncSecurityHandlerFactory(None)
     yield security_handler_factory
