+++ tests/fixtures/schema_references/openapi.yaml
@@ -1,0 +1,19 @@
+openapi: 3.0.0
+info:
+  title: '{{title}}'
+  version: '1.0'
+servers:
+  - url: /v1.0
+paths:
+ /pets:
+    get:
+      summary: List all pets
+      tags:
+        - pets
+      responses:
+        '200':
+          description: An paged array of pets
+          content:
+            application/json:
+              schema:
+                $ref: "#/components/schemas/Pets"
