+++ tests/fixtures/schema_references/swagger.yaml
@@ -1,0 +1,9 @@
+swagger: "2.0"
+
+info:
+  title: "{{title}}"
+  version: "1.0"
+
+basePath: /v1.0
+
+paths:
